requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
line-bot-sdk==1.8.0
psycopg2-binary==2.7.5
